System changes (SSL, Environment vars, help files)
Enable environment variable overrides to pass in containers and override the server configuration. Create default SSL if missing in conf folder. In bind mounts it will override the container location where SSL certificates are located. Would require a container to include copy SSL and configuration files to container or bind mount with SSL and configuration. Optional for environment variables in the container build to set any setting on the web services.